Install from multiple remotes
Loop over the words in CONAN_REMOTE_URL and CONAN_REMOTE_URL to add
multiple remote URLs. The variables should be set in
poky_build_folder/conf/local.conf

```
 CONAN_REMOTE_NAME = "conan-yocto conan-other"

 CONAN_REMOTE_URL = "\
                https://localhost:8081/artifactory/api/conan/<repository> \
                https://localhost:8081/artifactory/api/conan/<other-repository> \
 "
```
where the first URL in CONAN_REMOTE_URL corresponds to the first name in
CONAN_REMOTE_NAME, the second to the second etc. If the number of URLs
does not equal the number of names an error will be thrown.

The `--remote` option has been removed as conan will automatically try
all the remote URLs.

Note it assumes the CONAN_USER and CONAN_PASSWORD is the same for all
remote URLs.

resolves #14